+++ src/collector/tmpfs.rs
@@ -5,6 +5,7 @@ use crate::collector::{CollectionOutcome, CollectionStatus, ProbeOutcome};
 use crate::fs::FsRoots;
 use crate::output::OutputDir;
 
+#[derive(Clone)]
 pub struct Tmpfs;
 
 #[derive(Serialize)]